rename _test dir to _tests
